setup views - design home page
