+++ .env
@@ -1,2 +1,2 @@
 DB_URL = mongodb+srv://dinvvan:Chichichichi22@cluster0-rtonb.gcp.mongodb.net/test?retryWrites=true&w=majority
-PORT = 3001
+PORT = 3000
