+++ public/css/style.css
@@ -1,0 +1,86 @@
+@import url('https://fonts.googleapis.com/css?family=Indie+Flower');
+
+#main {
+  width: 60% !important;
+  margin: auto;
+  background-color: #0f0f0f;
+}
+
+html, body {
+  width: 100%;
+  height: 100%;
+  font-family: 'Indie Flower';
+  background-color: #171717;
+  font-size: 20 px;
+  color: #222;
+}
+
+#main a {
+  color: white !important;
+}
+
+.navbar {
+  background-color: rgba(0,0,0,.85) !important; /*độ trong*/
+}
+
+.navbar li a {
+  color: white !important;
+}
+
+.navbar li a:hover {
+  color: black !important;
+}
+
+.carousel-inner img {
+  width: 12.5%;
+}
+
+#slides {
+  padding: 1rem;
+}
+
+/*container*/
+
+.left-movies, .right-movies {
+  float: left;
+}
+
+.left-movies {
+  padding-top: 15px;
+}
+
+.title-list {
+  text-transform: uppercase;
+  color: white;
+  font-size: 25px;
+}
+
+.readmore {
+  float: right;
+  font-size: 15px;
+  position: relative;
+  top: 12px;
+}
+
+.movies-item {
+  padding-bottom: 10px;
+  padding-right: 5px !important
+}
+
+
+/*footer*/
+footer {
+  color: white;
+}
+
+.footer-header, .footer-bottom {
+  background-color: #159fff;
+  height: 2px;
+  opacity: .8;
+  margin: 5px;
+}
+
+.footer-header {
+  width: 75%;
+  margin: 20px auto;
+}
