+++ src/controllers/UserController.js
@@ -2,42 +2,42 @@ const User = require('../models/UserModels')
 const bcrypt = require('bcrypt')
 
 exports.register = function(req, res, next) {
-    User.findOne({email: req.body.email}, (err, user) => {
-        if (user = null) { //Kiểm tra xem email đã được sử dụng chưa
-            bcrypt.hash(req.body.password, 10, function (err, hash) { //Mã hóa mật khẩu trước khi lưu vào db
-                if (err) {return next(err);}
-                const user = new User(req.body)
-                user.role = ['customer'] //sau khi register thì role auto là customer
-                user.password = hash
-                user.password_confirm = hash
-                user.save((err, result) => {
-                    if (err) {return res.json({err})}
-                    res.json({user: result})
-                })
-            })
-        } else {
-            res.json({err: 'Email has been used'})
-        }
-    })
+  User.findOne({email: req.body.email}, (err, user) => {
+    if (user = null) { //Kiểm tra xem email đã được sử dụng chưa
+      bcrypt.hash(req.body.password, 10, function (err, hash) { //Mã hóa mật khẩu trước khi lưu vào db
+        if (err) {return next(err);}
+        const user = new User(req.body)
+        user.role = ['customer'] //sau khi register thì role auto là customer
+        user.password = hash
+        user.password_confirm = hash
+        user.save((err, result) => {
+            if (err) {return res.json({err})}
+            res.json({user: result})
+        })
+      })
+    } else {
+      res.json({err: 'Email has been used'})
+    }
+  })
 }
 
 exports.login = function (req, res) {
-    User.findOne({email: req.body.email}).exec(function(err, user) {
-        if (err) {
-            return res.json({err})
-        } else if (!user) {
-            return res.json({err: 'Username and Password are incorrect'})
-        }
-        bcrypt.compare(req.body.password, user.password, (err, result) => {
-            if(result === true){
-                req.session.user = user
-                res.json({
-                    user: user,
-                    "login": "sucess"
-                })
-            } else {
-                return res.json({err: 'Username and Password are incorrect!'})
-            }
+  User.findOne({email: req.body.email}).exec(function(err, user) {
+    if (err) {
+      return res.json({err})
+    } else if (!user) {
+      return res.json({err: 'Username and Password are incorrect'})
+    }
+    bcrypt.compare(req.body.password, user.password, (err, result) => {
+      if(result === true){
+        req.session.user = user
+        res.json({
+            user: user,
+            "login": "sucess"
         })
+      } else {
+        return res.json({err: 'Username and Password are incorrect!'})
+      }
     })
+  })
 }
